Exercise 4.21

+++ controllers/blogs.js
@@ -8,23 +8,21 @@ blogsRouter.get('/', async (request, response) => {
     response.json(blogs)
 })
 
-blogsRouter.get('/:id', (request, response, next) => {
-    Blog
-        .findById(request.params.id)
-        .then(blog => {
-            if (blog) {
-                response.json(blog)
-            }
-            else {
-                response.status(404).end()
-            }
-        })
-        .catch(error => next(error))
+blogsRouter.get('/:id', async (request, response, next) => {
+    const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1 })
+
+    if (blog) {
+        response.json(blog)
+    }
+    else {
+        response.status(404).end()
+    }
 })
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log('decptedToken',decodedToken)
 
     if (!decodedToken.id) {
         return response.status(401).json({ error: 'token invalid' })
@@ -47,8 +45,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'token invalid' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
